+++ src/tasks/batch_integration/control_methods/random_embed_cell/script.py
@@ -17,11 +17,6 @@ meta = {
 
 ## VIASH END
 
-with open(meta['config'], 'r', encoding="utf8") as file:
-    config = yaml.safe_load(file)
-
-output_type = config["functionality"]["info"]["subtype"]
-
 print('Read input', flush=True)
 input = ad.read_h5ad(par['input'])
 
@@ -32,6 +27,5 @@ input.obsm['X_emb'] = OneHotEncoder().fit_transform(
 )
 
 print("Store outputs", flush=True)
-input.uns['output_type'] = output_type
 input.uns['method_id'] = meta['functionality_name']
 input.write_h5ad(par['output'], compression='gzip')