+++ src/tasks/batch_integration/control_methods/random_embed_cell_jitter/script.py
@@ -20,11 +20,6 @@ meta = {
 
 ## VIASH END
 
-with open(meta['config'], 'r', encoding="utf8") as file:
-    config = yaml.safe_load(file)
-
-output_type = config["functionality"]["info"]["subtype"]
-
 print('Read input', flush=True)
 input = ad.read_h5ad(par['input'])
 
@@ -37,6 +32,5 @@ embedding = OneHotEncoder().fit_transform(
 input.obsm['X_emb'] = csr_matrix(embedding + np.random.uniform(-1 * par['jitter'], par['jitter'], embedding.shape))
 
 print("Store outputs", flush=True)
-input.uns['output_type'] = output_type
 input.uns['method_id'] = meta['functionality_name']
 input.write_h5ad(par['output'], compression='gzip')