+++ src/tasks/batch_integration/methods/scvi/script.py
@@ -14,11 +14,6 @@ meta = {
 }
 ## VIASH END
 
-with open(meta['config'], 'r', encoding="utf8") as file:
-    config = yaml.safe_load(file)
-
-output_type = config["functionality"]["info"]["subtype"]
-
 print('Read input', flush=True)
 adata = ad.read_h5ad(par['input'])
 
@@ -32,6 +27,5 @@ adata = scvi(adata, batch='batch')
 del adata.X
 
 print("Store outputs", flush=True)
-adata.uns['output_type'] = output_type
 adata.uns['method_id'] = meta['functionality_name']
 adata.write_h5ad(par['output'], compression='gzip')