+++ CHANGELOG.md
@@ -108,6 +108,8 @@
 
 * `resources_test/label_projection/pancreas` with `src/tasks/label_projection/resources_test_scripts/pancreas.sh`.
 
+* `workflows/run`: Added nf-tower test script. (PR #205)
+
 ### V1 MIGRATION
 
 * Removed the separate subtask specific subfolders. The "subtask" is added to the config.
@@ -140,6 +142,10 @@
 
 * `metrics/pcr`: Migrated from v1 embedding.
 
+### MINOR CHANGES
+
+* Removed the `.uns["output_type"]` field from output anndata in methods and control methods. (PR #205)
+
 ## label_projection
 
 ### NEW FUNCTIONALITY
@@ -151,6 +157,8 @@
 * `process_dataset`: Added a component for processing common datasets into task-ready dataset objects.
 
 * `resources_test/label_projection/pancreas` with `src/tasks/label_projection/resources_test_scripts/pancreas.sh`.
+
+* * `workflows/run`: Added nf-tower test script. (PR #205)
 
 ### V1 MIGRATION
 
@@ -189,6 +197,8 @@
 * `process_dataset`: Added a component for processing common datasets into task-ready dataset objects.
 
 * `resources_test/denoising/pancreas` with `src/tasks/denoising/resources_test_scripts/pancreas.sh`.
+  
+* `workflows/run`: Added nf-tower test script. (PR #205)
 
 ### V1 MIGRATION
 
@@ -230,6 +240,8 @@
 * `resources_test/dimensionality_reduction/pancreas` with `src/tasks/dimensionality_reduction/resources_test_scripts/pancreas.sh`.
 
 * Added `variant` key to config files to store variants (different input parameters) of every component.
+  
+* `workflows/run`: Added nf-tower test script. (PR #205)
 
 ### V1 migration
 * `control_methods/true_features`: Migrated from v1. Extracted from baseline method `True Features`.